refactor: useFlip 커스텀훅 분리
