+++ frontend/src/domains/preferRestaurant/components/PreferRestaurantList.tsx
@@ -1,7 +1,8 @@
 import { Restaurant } from '@apis/restaurant';
 
+import { useFlip } from '@hooks/useFlip';
+
 import styled from '@emotion/styled';
-import { useEffect, useRef } from 'react';
 
 import { usePreferRestaurantContext } from '../context/PreferRestaurantProvider';
 
@@ -9,47 +10,8 @@ import PreferRestaurantItem from './PreferRestaurantItem';
 
 function PreferRestaurantList() {
   const { restaurantList } = usePreferRestaurantContext();
-  const itemRefs = useRef(new Map<string, HTMLDivElement>());
-  const prevRects = useRef(new Map<string, DOMRect>());
 
-  const recordPositions = () => {
-    restaurantList.forEach(item => {
-      const el = itemRefs.current.get(item.id);
-      if (el) {
-        prevRects.current.set(item.id, el.getBoundingClientRect());
-      }
-    });
-  };
-
-  recordPositions();
-
-  const playFLIPAnimations = () => {
-    restaurantList.forEach(item => {
-      const el = itemRefs.current.get(item.id);
-      const prevRect = prevRects.current.get(item.id);
-      if (!el || !prevRect) return;
-
-      const newRect = el.getBoundingClientRect();
-      const dx = prevRect.left - newRect.left;
-      const dy = prevRect.top - newRect.top;
-
-      if (dx !== 0 || dy !== 0) {
-        el.style.transform = `translate(${dx}px, ${dy}px)`;
-        el.style.transition = 'transform 0s';
-
-        requestAnimationFrame(() => {
-          el.style.transition = 'transform 500ms ease';
-          el.style.transform = 'translate(0, 0)';
-        });
-      }
-    });
-  };
-
-  useEffect(() => {
-    requestAnimationFrame(() => {
-      playFLIPAnimations();
-    });
-  }, [restaurantList]);
+  const { itemRefs } = useFlip(restaurantList);
 
   return (
     <S.Container>
